# OpenAPI.yaml
openapi: 3.0.0
info:
  title: store
  version: '1.0'
  description: A simple store
  contact:
    name: SecuritEase Dev
    url: 'https://www.securitease.com'
    email: internal@securitease.com
  termsOfService: 'https://www.securitease.com'
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.htm'
servers:
  - url: 'http://localhost:8080'
    description: dev
paths:
  /order:
    get:
      summary: Get Orders
      tags:
        - store
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  page:
                    $ref: '#/components/schemas/PageMetadata'
      operationId: getOrders
      description: Get orders, paginated
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/sort'
    post:
      summary: Create Order
      tags:
        - store
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request - customer or product ID not found
      operationId: createOrder
      description: Create a new order
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                customerId:
                  type: integer
                  format: int64
                productIds:
                  type: array
                  items:
                    type: integer
                    format: int64
  /order/{id}:
    get:
      summary: Get Order by ID
      tags:
        - store
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Not Found
      operationId: getOrderById
      description: Get a specific order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
  /customer:
    get:
      summary: Get Customers
      tags:
        - store
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Customer'
                  page:
                    $ref: '#/components/schemas/PageMetadata'
      operationId: getCustomers
      description: Get customers, paginated, optionally filtered by a substring match against one of the words in their name
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
          description: Substring to match against a word in the customer's name
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/sort'
    post:
      summary: Create Customer
      tags:
        - store
      responses:
        '201':
          description: Created
      operationId: createCustomer
      description: Create a new customer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
  /products:
    get:
      summary: Get Products
      tags:
        - store
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  page:
                    $ref: '#/components/schemas/PageMetadata'
      operationId: getProducts
      description: Get products, paginated
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/sort'
    post:
      summary: Create Product
      tags:
        - store
      responses:
        '201':
          description: Created
      operationId: createProduct
      description: Create a new product
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
  /products/{id}:
    get:
      summary: Get Product by ID
      tags:
        - store
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Not Found
      operationId: getProductById
      description: Get a specific product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
components:
  parameters:
    page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 0
      description: Zero-based page index
    size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 20
      description: Page size (capped at 100)
    sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
      description: 'Sort order, e.g. "id,desc"'
  schemas:
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
        number:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
    Order:
      title: Order
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        customer:
          type: object
          properties:
            id:
              type: integer
              format: int64
            description:
              type: string
        products:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                format: int64
              description:
                type: string
    Product:
      title: Product
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        orders:
          type: array
          items:
            type: integer
            format: int64
    Customer:
      title: Customer
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        orders:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                format: int64
              description:
                type: string
                x-stoplight:
                  id: mq08x82g95zl9
tags:
  - name: store
    description: store operations